CRUD con pequeñas validaciones completado, realizar peubas de ambos

+++ controllers/asset_controller.js
@@ -1,5 +1,7 @@
 const model = require("../models/asset");
 
+const NotFoundError = require("../errors/NotFoundError");
+
 //--FILTRAO GENERAL--
 const getAllAssets = async (req, res) => {
   const assets = await model.getAllAssetsModel();
@@ -7,30 +9,42 @@ const getAllAssets = async (req, res) => {
   res.json({ data: assets });
 };
 
-const getAssetByEmployeeId = async (req, res)=>{
-  const assets = await model.getAssetByEmployeeIdModel();
-  res.json({data:assets})
-}
+const getAssetByEmployeeId = async (req, res) => {
+  const userId = req.params.id;
+  if (!userId) {
+    getAllAssets();
+  } else {
+    const assets = await model.getAssetByEmployeeIdModel(userId);
+    res.json({ data: assets });
+  }
+};
 
 //--TRAER ASSET POR SU ID--
 const getAssetById = async (req, res, next) => {
   try {
     const assetId = req.params.aid;
     const asset = await model.getAssetByIdModel(assetId);
+    if (!asset) throw new NotFoundError("El Articulo no existe");
     res.json({ data: asset });
   } catch (error) {
-    next(error)
+    next(error);
   }
 };
 
 //--CREAR ASSET--
 const createAsset = async (req, res) => {
+  //Validacion de resultados de express validator: Me surgian muchos errores mandando a un archivo aparte
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(400).json({ errors: errors.array() });
+  }
+
   try {
-    const values = {...req.body}; //--Traemos valores del cuerpo
+    const values = { ...req.body }; //--Traemos valores del cuerpo
     const result = await model.createAssetModel(values);
     res.json({ data: result, messaje: "Asset Creado" });
   } catch (error) {
-    next(error)
+    next(error);
   }
 };
 
@@ -38,11 +52,16 @@ const createAsset = async (req, res) => {
 const updateAsset = async (req, res) => {
   try {
     const assetId = req.params.aid;
-    const values = {...req.body};
-    const result = await model.updateAssetModel(assetId,values)
-    res.json({result, messaje: `El asset con el id ${assetId} se actualizó exitosamente`})
+    const asset = await model.getAssetByIdModel(assetId);
+    if (!asset) throw new NotFoundError("El Articulo no existe");
+    const values = { ...req.body };
+    const result = await model.updateAssetModel(asset, values);
+    res.json({
+      result,
+      messaje: `El asset con el id ${assetId} se actualizó exitosamente`,
+    });
   } catch (error) {
-    next(error)
+    next(error);
   }
 };
 
@@ -50,10 +69,12 @@ const updateAsset = async (req, res) => {
 const deleteAseet = async (req, res) => {
   try {
     const assetId = req.params.aid;
+    const asset = await model.getAssetByIdModel(assetId);
+    if (!asset) throw new NotFoundError("El Articulo no existe");
     await model.deleteAssetModel(assetId);
-    res.json({ messaje: `Usuario con el Id: ${assetId} eliminado` });
+    res.json({ messaje: `Articulo con el Id: ${assetId} eliminado` });
   } catch (error) {
-    next(error)
+    next(error);
   }
 };
 module.exports = {
